Forslag på å ikke kjøre Kafka lokalt i test
I denne mer lettbente løsningen mocker vi ut Kafka-produsenten
så vi slipper å kjøre Kafka lokalt.
Løsningen gir mindre kode og slipper styret med å starte opp
Kafka og få det til å fungere stabilt.

Når vi starter applikasjonen lokalt stubbes Kafka-produsenten.
