+++ src/Application.kt
@@ -3,7 +3,7 @@ package no.nav.rekrutteringsbistand.statistikk
 import io.ktor.auth.Authentication
 import io.ktor.util.KtorExperimentalAPI
 import no.nav.rekrutteringsbistand.statistikk.db.Database
-import no.nav.rekrutteringsbistand.statistikk.kafka.DatavarehusKafkaProducer
+import no.nav.rekrutteringsbistand.statistikk.kafka.DatavarehusKafkaProducerImpl
 import no.nav.security.token.support.ktor.tokenValidationSupport
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
@@ -22,7 +22,7 @@ fun main() {
     val applicationEngine = lagApplicationEngine(
         database = database,
         tokenValidationConfig = tokenValidationConfig,
-        datavarehusKafkaProducer = DatavarehusKafkaProducer("todo") // TODO
+        datavarehusKafkaProducer = DatavarehusKafkaProducerImpl("todo") // TODO
     )
     applicationEngine.start()
     log.info("Applikasjon startet i miljø: ${Cluster.current}")
