--- test/KafkaConsumerUtils.kt
@@ -1,21 +1,0 @@
-import org.apache.kafka.clients.CommonClientConfigs
-import org.apache.kafka.clients.consumer.ConsumerConfig
-import org.apache.kafka.clients.consumer.KafkaConsumer
-import org.apache.kafka.common.serialization.StringDeserializer
-import java.util.*
-
-class KafkaConsumerUtils {
-
-    companion object {
-        fun opprettConsumer(bootstrapServers: String): KafkaConsumer<String, String> {
-            val consumerConfig: Properties = Properties().apply {
-                put(CommonClientConfigs.BOOTSTRAP_SERVERS_CONFIG, bootstrapServers)
-                put(ConsumerConfig.AUTO_OFFSET_RESET_CONFIG, "earliest")
-                put(ConsumerConfig.GROUP_ID_CONFIG, "mingroupid")
-                put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer::class.java)
-                put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer::class.java)
-            }
-            return KafkaConsumer(consumerConfig)
-        }
-    }
-}
