+++ test/LagreStatistikkTest.kt
@@ -15,10 +15,7 @@ import io.ktor.http.content.TextContent
 import io.ktor.util.KtorExperimentalAPI
 import kotlinx.coroutines.runBlocking
 import db.TestDatabase
-import no.nav.common.KafkaEnvironment
-import no.nav.rekrutteringsbistand.statistikk.log
 import org.junit.Test
-import java.time.Duration
 import java.time.LocalDateTime
 import java.time.temporal.ChronoUnit
 
@@ -35,10 +32,9 @@ class LagreStatistikkTest {
     companion object {
         private val database = TestDatabase()
         private val port = randomPort()
-        private val lokalKafka = KafkaEnvironment()
 
         init {
-            start(database, port, lokalKafka)
+            start(database, port)
         }
     }
 
@@ -60,8 +56,6 @@ class LagreStatistikkTest {
             assertThat(utfall.stillingsId).isEqualTo(kandidatutfallTilLagring[index].stillingsId)
             assertThat(utfall.tidspunkt.truncatedTo(ChronoUnit.MINUTES)).isEqualTo(LocalDateTime.now().truncatedTo(ChronoUnit.MINUTES))
         }
-
-        lokalKafka.tearDown()
     }
 
     @Test
@@ -69,33 +63,5 @@ class LagreStatistikkTest {
         val uinnloggaClient = HttpClient(Apache)
         val response: HttpResponse = uinnloggaClient.post("$basePath/kandidatutfall")
         assertThat(response.status).isEqualTo(HttpStatusCode.Unauthorized)
-
-        lokalKafka.tearDown()
     }
-
-//    @Test
-//    fun `POST til kandidatutfall skal produsere melding på Kafka-topic`() = runBlocking {
-//        lokalKafka.adminClient?.deleteTopics(listOf(DatavarehusKafkaProducer.TOPIC))
-//
-//        val consumer = KafkaConsumerUtils.opprettConsumer(lokalKafka.brokersURL)
-//        consumer.subscribe(listOf(DatavarehusKafkaProducer.TOPIC))
-//
-//
-//        val kandidatutfallTilLagring = listOf(etKandidatutfall, etKandidatutfall)
-//
-//        val response: HttpResponse = client.post("$basePath/kandidatutfall") {
-//            body = TextContent(tilJson(kandidatutfallTilLagring), ContentType.Application.Json)
-//        }
-//
-//        consumer.poll(Duration.ofSeconds(5))
-//            .map { melding -> jacksonObjectMapper().readValue<OpprettKandidatutfall>(melding.value()) }
-//            .forEachIndexed { index, melding ->
-//                assertThat(melding.aktørId).isEqualTo(kandidatutfallTilLagring[index].aktørId)
-//                assertThat(melding.utfall).isEqualTo(kandidatutfallTilLagring[index].utfall)
-//                assertThat(melding.navIdent).isEqualTo(kandidatutfallTilLagring[index].navIdent)
-//                assertThat(melding.navKontor).isEqualTo(kandidatutfallTilLagring[index].navKontor)
-//                assertThat(melding.kandidatlisteId).isEqualTo(kandidatutfallTilLagring[index].kandidatlisteId)
-//                assertThat(melding.stillingsId).isEqualTo(kandidatutfallTilLagring[index].stillingsId)
-//            }
-//    }
 }
