+++ test/kafka/DatavarehusKafkaProducerStub.kt
@@ -1,0 +1,12 @@
+package kafka
+
+import no.nav.rekrutteringsbistand.statistikk.kafka.DatavarehusKafkaProducer
+import no.nav.rekrutteringsbistand.statistikk.kandidatutfall.OpprettKandidatutfall
+import no.nav.rekrutteringsbistand.statistikk.log
+
+class DatavarehusKafkaProducerStub: DatavarehusKafkaProducer {
+
+    override fun send(kandidatutfall: OpprettKandidatutfall) {
+        log.info("Stubber sending av kandidatutfall til Datavarehus. Kandidatutfall: $kandidatutfall")
+    }
+}
